infra/requirements.txt
# AWS CDK dependencies for Pantry Pirate Radio infrastructure
aws-cdk-lib>=2.120.0
constructs>=10.0.0

# Testing
pytest>=7.4.0
